Use `GradlePropertiesController` instead of `StartParameterInternal` for the DCL flag
