+++ subprojects/core/src/main/java/org/gradle/api/internal/StartParameterInternal.java
@@ -42,7 +42,6 @@ public class StartParameterInternal extends StartParameter {
     private boolean configurationCacheParallel;
     private boolean configurationCacheRecreateCache;
     private boolean configurationCacheQuiet;
-    private boolean kotlinDslDclEnabled;
     private boolean searchUpwards = true;
     private boolean useEmptySettings = false;
     private Duration continuousBuildQuietPeriod = Duration.ofMillis(250);
@@ -80,7 +79,6 @@ public class StartParameterInternal extends StartParameter {
         p.configurationCacheParallel = configurationCacheParallel;
         p.configurationCacheRecreateCache = configurationCacheRecreateCache;
         p.configurationCacheQuiet = configurationCacheQuiet;
-        p.kotlinDslDclEnabled = kotlinDslDclEnabled;
         p.searchUpwards = searchUpwards;
         p.useEmptySettings = useEmptySettings;
         return p;
@@ -248,13 +246,5 @@ public class StartParameterInternal extends StartParameter {
 
     public void setPropertyUpgradeReportEnabled(boolean propertyUpgradeReportEnabled) {
         this.propertyUpgradeReportEnabled = propertyUpgradeReportEnabled;
-    }
-
-    public boolean isKotlinDslDclEnabled() {
-        return kotlinDslDclEnabled;
-    }
-
-    public void setKotlinDslDclEnabled(boolean kotlinDslDclEnabled) {
-        this.kotlinDslDclEnabled = kotlinDslDclEnabled;
     }
 }
