+++ subprojects/core/src/main/java/org/gradle/initialization/StartParameterBuildOptions.java
@@ -82,7 +82,6 @@ public class StartParameterBuildOptions extends BuildOptionSet<StartParameterInt
         options.add(new ConfigurationCacheParallelOption());
         options.add(new ConfigurationCacheRecreateOption());
         options.add(new ConfigurationCacheQuietOption());
-        options.add(new KotlinDslDslEnabledOption());
         options.add(new IsolatedProjectsOption());
         options.add(new PropertyUpgradeReportOption());
         StartParameterBuildOptions.options = Collections.unmodifiableList(options);
@@ -634,20 +633,6 @@ public class StartParameterBuildOptions extends BuildOptionSet<StartParameterInt
         @Override
         public void applyTo(boolean value, StartParameterInternal settings, Origin origin) {
             settings.setConfigurationCacheQuiet(value);
-        }
-    }
-
-    public static class KotlinDslDslEnabledOption extends BooleanBuildOption<StartParameterInternal> {
-
-        public static final String PROPERTY_NAME = "org.gradle.kotlin.dsl.dcl";
-
-        public KotlinDslDslEnabledOption() {
-            super(PROPERTY_NAME);
-        }
-
-        @Override
-        public void applyTo(boolean value, StartParameterInternal settings, Origin origin) {
-            settings.setKotlinDslDclEnabled(value);
         }
     }
